+++ packages/webpack/src/loader-options.ts
@@ -7,6 +7,7 @@
 
 /* eslint-disable @typescript-eslint/no-non-null-assertion */
 import { LoaderContext, WebpackError } from "webpack";
+import { DEFAULTS } from "./options";
 import Upath from "./Upath";
 
 export interface PluginArguments {
@@ -33,7 +34,8 @@ export const getLoaderOptions = (loader: LoaderContext<PluginOptions>): PluginOp
         ...options,
         ...(loader._module && loader._module.addWarning && { warn: (err: WebpackError) => loader._module!.addWarning(err) }),
         ...(loader._module && loader._module.addError && { error: (err: WebpackError) => loader._module!.addError(err) }),
-        config: Upath.resolve(options.config),
+        config: Upath.resolve(options.config ?? DEFAULTS.config),
+        transpileOnly: options.transpileOnly ?? false,
         webpackTarget: options.webpackTarget ?? "*",
     };
     return result;