+++ Release-Notes.md
@@ -20,7 +20,8 @@ Release notes follow the [keep a changelog](https://keepachangelog.com/en/1.0.0/
 
 ### Fixed
 
-- Corrects bundling of websmith-core in websmith-compiler and websmith-webpack
+- Corrects bundling of websmith-core in websmith-compiler and websmith-webpack.
+- Webpack module correctly processes webpack config options.
 
 ## [0.3.0] - 2022-11-17